fix: use `touchNode` correctly (passing the full node) and use Gatsby `reporter` for logging (#3)
Between Gatsby v3 and v4 the API of `touchNode` changed, see [Migrating from v3 to v4](https://www.gatsbyjs.com/docs/reference/release-notes/migrating-from-v3-to-v4/#change-arguments-passed-to-touchnode-action).

Previously, `touchNode` accepted an object with just the `nodeId`.

Now, `touchNode` expects the _full node_ to be passed as argument. We are using the `getNode()` API as suggested by the migration guide when working with cached node ids.